modif jardin + retention

+++ Module_Persistence/src/main/java/com/fr/adaming/entity/Jardin.java
@@ -38,7 +38,7 @@ public class Jardin {
 	@Column
 	private Float ProfSol;
 	@Column
-	private float reserveUtile;
+	private double reserveUtile;
 	@Column
 	private boolean estArroser ;
